+++ core/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Print inverted question mark for non-printable characters.
 - Remove pre-fill bar from text rendering functions.  [#1173]
 - Display coin name when signing or verifying messages.  [#1159]
+- Allow spending coins from Bitcoin paths if the coin has implemented strong replay protection via `SIGHASH_FORKID`.  [#1188]
 
 ### Deprecated
 
@@ -269,3 +270,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [#1159]: https://github.com/trezor/trezor-firmware/issues/1159
 [#1165]: https://github.com/trezor/trezor-firmware/pull/1165
 [#1173]: https://github.com/trezor/trezor-firmware/pull/1173
+[#1188]: https://github.com/trezor/trezor-firmware/issues/1188